backend/requirements.txt
Django==4.2.30
djangorestframework==3.16.1
django-cors-headers==4.9.0
python-decouple==3.8
requests==2.32.5
gunicorn==23.0.0
whitenoise==6.11.0
